requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url ./requirements.in
#
colorama==0.4.6
    # via qrcode
numpy==1.24.3
    # via opencv-python-headless
opencv-python-headless==4.5.5.64
    # via -r ./requirements.in
pillow==9.5.0
    # via
    #   -r ./requirements.in
    #   qrcode
pypng==0.20220715.0
    # via qrcode
pyside6==6.5.0
    # via -r ./requirements.in
pyside6-addons==6.5.0
    # via pyside6
pyside6-essentials==6.5.0
    # via pyside6
qrcode[pil]==7.4.2
    # via -r ./requirements.in
shiboken6==6.5.0
    # via pyside6
typing-extensions==4.5.0
    # via qrcode
zbar-lite==0.23.93
    # via -r ./requirements.in
